Fix performance regression in the IDE diagnostic analyzer service.
Currently we maintain a cache of CompilationWithAnalyzers for projects being actively analyzed. This is causing the VM to rise and also forcing GC to kick in at different phases causing regressions at various scenarios. The fix is to just hold onto a single CompilationCompilationWithAnalyzers instance for a project being analyzed.